init commit for gf2 polish and shorten tree bench
