+++ arith/gf2/src/gf2x128/avx.rs
@@ -4,9 +4,9 @@ use std::{
     ops::{Add, AddAssign, Mul, MulAssign, Neg, Sub, SubAssign},
 };
 
-use arith::{Field, FieldSerde, FieldSerdeResult, SimdField};
+use arith::{Field, FieldSerde, FieldSerdeResult};
 
-use crate::{GF2x64, GF2};
+use crate::GF2;
 
 #[derive(Debug, Clone, Copy)]
 pub struct AVXGF2x128 {
@@ -37,7 +37,7 @@ impl FieldSerde for AVXGF2x128 {
 }
 
 impl Field for AVXGF2x128 {
-    const NAME: &'static str = "Galios Field 2 SIMD 128";
+    const NAME: &'static str = "AVX Galois Field 2 SIMD 128";
 
     const SIZE: usize = 128 / 8;
 
@@ -301,43 +301,5 @@ impl From<GF2> for AVXGF2x128 {
         } else {
             AVXGF2x128::ONE
         }
-    }
-}
-
-impl SimdField for AVXGF2x128 {
-    type Scalar = GF2;
-
-    const PACK_SIZE: usize = 128;
-
-    #[inline(always)]
-    fn scale(&self, challenge: &Self::Scalar) -> Self {
-        if challenge.v == 0 {
-            Self::ZERO
-        } else {
-            *self
-        }
-    }
-
-    #[inline(always)]
-    fn pack(base_vec: &[Self::Scalar]) -> Self {
-        assert_eq!(base_vec.len(), Self::PACK_SIZE);
-        let mut packed_to_gf2x64 = [GF2x64::ZERO; Self::PACK_SIZE / GF2x64::PACK_SIZE];
-        packed_to_gf2x64
-            .iter_mut()
-            .zip(base_vec.chunks(GF2x64::PACK_SIZE))
-            .for_each(|(gf2x64, pack)| *gf2x64 = GF2x64::pack(pack));
-
-        unsafe { transmute(packed_to_gf2x64) }
-    }
-
-    #[inline(always)]
-    fn unpack(&self) -> Vec<Self::Scalar> {
-        let packed_to_gf2x64: [GF2x64; Self::PACK_SIZE / GF2x64::PACK_SIZE] =
-            unsafe { transmute(*self) };
-
-        packed_to_gf2x64
-            .iter()
-            .flat_map(|packed| packed.unpack())
-            .collect()
     }
 }
