+++ arith/gf2_128/src/gf2_ext128x8/neon.rs
@@ -55,7 +55,7 @@ impl FieldSerde for NeonGF2_128x8 {
 }
 
 impl Field for NeonGF2_128x8 {
-    const NAME: &'static str = "Neon Galios Field 2 128x8";
+    const NAME: &'static str = "Neon Galois Field 2^128 SIMD 8";
 
     const SIZE: usize = 16 * 8;
