+++ tree/Cargo.toml
@@ -17,7 +17,6 @@ rayon.workspace = true
 criterion.workspace = true
 gf2 = { path = "../arith/gf2" }
 gf2_128 = { path = "../arith/gf2_128" }
-mersenne31 = { path = "../arith/mersenne31" }
 tynm.workspace = true
 
 # p3-baby-bear.workspace = true
